PostView2: Continued implementation of data Filter feature.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/PostView2@8555 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ PostView2/GLWidgetManager.h
@@ -20,7 +20,7 @@ public:
 
 	void AddWidget(GLWidget* pw);
 	void RemoveWidget(GLWidget* pw);
-	int Widgets() { return m_Widget.size(); }
+	int Widgets() { return (int)m_Widget.size(); }
 
 	GLWidget* operator [] (int i) { return m_Widget[i]; }
